requirements.txt
matplotlib==3.10.0
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
timm==1.0.14
torch==2.0.1+cu117
torchvision==0.15.2+cu117
transformers==4.47.1
